+++ src/sessions/user.session.js
@@ -17,10 +17,6 @@ export const removeUser = async (socket) => {
 };
 
 export const getUserBySocket = async (socket) => {
-  console.log('usersocket');
-  console.log(userSessions);
-  console.log('socket');
-  console.log(socket);
   const user = userSessions.find((user) => user.socket === socket);
   if (!user) {
     console.error('User not found');
